pyproject.toml
[tool.poetry]
name = "kt-datacards"
version = "0.1.0"
description = "Kill Team datacard extraction and processing for Tabletop Simulator"
authors = ["Wen-Qualtu"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pymupdf = "^1.23.8"
pillow = "^10.0.0"
pypdf2 = "^3.0.0"
pytesseract = "^0.3.10"
pyyaml = "^6.0"
pandas = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
